// apps/nativeapp/src/types/interface/slice.interface.ts
export interface RegisterFormSliceInitialState {
  form_id: string
  key: INTERVENTION_TYPE
  title: string
  intervention_date: number
  skip_intervention_form: boolean
  user_type: string
  project_id: string
  project_name: string
  site_id: string
  site_name: string
  can_be_entire_site: boolean
  entire_site_selected: boolean
  should_register_location: boolean
  location_type: 'Point' | 'Polygon'
  location_title: string
  coordinates: Array<number[]>
  preview_blank_polygon: boolean
  cover_image_url: string
  species_required: boolean
  is_multi_species: boolean
  species_count_required: boolean
  species_modal_message: string
  species_modal_unit: string
  species: string[]
  has_sample_trees: boolean
  tree_details_required: boolean
  tree_details: SampleTree[]
  form_details: MainForm[]
  meta_data: string
  additional_data: FormElement[]
  form_data: FormElement[]
  plantedSpecies: PlantedSpecies[]
  optionalLocation?: boolean
}

export interface QuaeBody {
  type: string
  priority: number
  nextStatus: INTERVENTION_STATUS
  p1Id?: string
  p2Id?: string
  p3Id?: string
}

export interface SampleTree {
  project_id?: string
  site_id?: string
  tree_id: string
  species_guid: string
  intervention_id: string
  sloc_id: string
  count: number
  latitude: number
  longitude: number
  device_latitude: number
  device_longitude: number
  location_accuracy: string
  image_url: string
  cdn_image_url: string
  specie_name: string
  local_name: string
  specie_diameter: number
  specie_height: number
  tag_id: string
  plantation_date: number
  status_complete: boolean
  location_id: string
  tree_type: 'sample' | 'single'
  parent_id: string
  additional_details: string
  app_meta_data: string
  hid: string,
  history: History[],
  status: INTERVENTION_STATUS
  remeasurement_dates: RemeasurementDate
  remeasurement_requires: boolean,
  is_alive: boolean,
  fix_required: FIX_REQUIRED
  image_data: {
    latitude: number
    longitude: number
    imageUrl: string
    cdnImageUrl: string
    currentloclat: number
    currentloclong: number
    isImageUploaded: boolean
    coordinateID: string
  },
}

export interface PlantedSpecies {
  id?: string,
  guid: string,
  scientificName: string,
  aliases: string,
  count: number,
  image: string
}
// ...
export interface RemeasurementDate {
  sampleTreeId: string
  created: number
  lastMeasurement: number
  remeasureBy: number
  nextMeasurement: number
}

export interface History {
  history_id: string
  eventName: string
  eventDate: number
  imageUrl: string
  cdnImageUrl: string
  diameter: number
  height: number
  additionalDetails: any
  appMetadata: string
  status: string
  statusReason: string
  dataStatus: 'SYNCED' |  "REMEASUREMENT_DATA_UPLOAD" | "REMEASUREMENT_EVENT_UPDATE" | "SKIP_REMEASUREMENT"
  parentId: string
  samplePlantLocationIndex: number
  lastScreen: string
}